+++ src/libs/bin/mod.rs
@@ -14,6 +14,7 @@ pub mod map_lib;
 pub mod math_lib;
 pub mod rand_lib;
 pub mod reg_lib;
+pub mod se_lib;
 pub mod string_lib;
 pub mod sys_lib;
 pub mod time_lib;